Update ZKInputs to incl Accumulated Fees of val 0
Fixes the case where there is a PoolTxs of type Exit with fee 0 for a
TokenID. Before this commit the Coordinator was not sending the
accumulated fee (which has value 0) to the Coordinator account for that
TokenID, with this commit the Coordinator always sends the accumulated
fee even when the value is 0 to match the hermez circom circuits
behaviour.

Also added a test to check the values and that also sends the proof to a
real proof server to check that can generate a valid proof.